+++ unnamed_engine/src/renderer/renderer.rs
@@ -78,7 +78,7 @@ impl Renderer {
     let camera_controller = CameraController::new(
       &device,
       4.0,
-      0.1,
+      0.4,
       viewport.config.width,
       viewport.config.height,
     );
@@ -103,7 +103,7 @@ impl Renderer {
   pub fn resize(&mut self, new_size: winit::dpi::PhysicalSize<u32>) {
     if new_size.width > 0 && new_size.height > 0 {
       self.viewport.resize(&self.device, new_size);
-      self.camera_controller.projection.resize(new_size.width, new_size.height);
+      self.camera_controller.resize(new_size.width, new_size.height);
 
       // Request a redraw just in case
       self.viewport.desc.window.request_redraw();